api_rest/requirements.txt
Django==1.11.11
django-filter==2.0.0
djangorestframework==3.9.0
Markdown==3.0.1
PyMySQL==0.9.2
request==2018.11.20
requests==2.20.0

